refactor: update post URL helper functions
