+++ src/components/Card.astro
@@ -1,9 +1,8 @@
 ---
 import type { CollectionEntry } from "astro:content";
 import { slugifyStr } from "@/utils/slugify";
-import { getPath } from "@/utils/getPath";
+import { getPostUrl } from "@/utils/postPaths";
 import Datetime from "./Datetime.astro";
-import config from "@/config";
 
 type Props = {
   variant?: "h2" | "h3";
@@ -11,14 +10,12 @@ type Props = {
 
 const { variant: Heading = "h2", id, data, filePath } = Astro.props;
 
-const locale = Astro.currentLocale ?? config.site.lang;
-
 const { title, description, ...props } = data;
 ---
 
 <li class="my-6">
   <a
-    href={getPath(id, filePath, true, locale)}
+    href={getPostUrl(id, filePath, Astro.currentLocale)}
     class:list={[
       "text-accent inline-block text-lg font-medium",
       "decoration-dashed underline-offset-4 hover:underline",
