+++ src/pages/posts/[...slug]/_components/AdjacentPostNav.astro
@@ -3,7 +3,7 @@ import { getCollection } from "astro:content";
 import IconArrowLeft from "@/assets/icons/IconArrowLeft.svg";
 import IconArrowRight from "@/assets/icons/IconArrowRight.svg";
 import { getSortedPosts } from "@/utils/getSortedPosts";
-import { getPath } from "@/utils/getPath";
+import { getPostUrl } from "@/utils/postPaths";
 import { useTranslations } from "@/i18n";
 import config from "@/config";
 
@@ -36,7 +36,7 @@ const t = useTranslations(locale);
   {
     prevPost && (
       <a
-        href={getPath(prevPost.id, prevPost.filePath, true, locale)}
+        href={getPostUrl(prevPost.id, prevPost.filePath, locale)}
         class="flex w-full gap-1 hover:opacity-75"
       >
         <IconArrowLeft class="inline-block flex-none rtl:rotate-180" />
@@ -50,7 +50,7 @@ const t = useTranslations(locale);
   {
     nextPost && (
       <a
-        href={getPath(nextPost.id, nextPost.filePath, true, locale)}
+        href={getPostUrl(nextPost.id, nextPost.filePath, locale)}
         class="flex w-full justify-end gap-1 text-end hover:opacity-75 sm:col-start-2"
       >
         <div>
