+++ src/pages/posts/[...slug]/index.png.ts
@@ -1,6 +1,6 @@
 import type { APIRoute } from "astro";
 import { getCollection, type CollectionEntry } from "astro:content";
-import { getPath } from "@/utils/getPath";
+import { getPostSlug } from "@/utils/postPaths";
 import { generateOgImageForPost } from "@/utils/og/renderOgImage";
 import config from "@/config";
 
@@ -14,7 +14,7 @@ export async function getStaticPaths() {
   );
 
   return posts.map(post => ({
-    params: { slug: getPath(post.id, post.filePath, false) },
+    params: { slug: getPostSlug(post.id, post.filePath) },
     props: post,
   }));
 }
