+++ src/pages/rss.xml.ts
@@ -1,20 +1,19 @@
 import rss from "@astrojs/rss";
 import { getCollection } from "astro:content";
 import { getSortedPosts } from "@/utils/getSortedPosts";
-import { getPath } from "@/utils/getPath";
+import { getPostUrl } from "@/utils/postPaths";
 import config from "@/config";
 
 export async function GET() {
   const posts = await getCollection("posts");
   const sortedPosts = getSortedPosts(posts);
-  const locale = config.site.lang;
 
   return rss({
     title: config.site.title,
     description: config.site.description,
     site: config.site.url,
     items: sortedPosts.map(({ data, id, filePath }) => ({
-      link: getPath(id, filePath, true, locale),
+      link: getPostUrl(id, filePath, config.site.lang),
       title: data.title,
       description: data.description,
       pubDate: new Date(data.modDatetime ?? data.pubDatetime),
